backend_bedrock/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
boto3==1.35.20
botocore==1.35.20
PyJWT==2.9.0

